+++ lib/src/cover_crypt/cover_crypt.dart
@@ -1,5 +1,6 @@
 import 'dart:typed_data';
 
+import 'cleartext_header.dart';
 import 'ffi.dart';
 
 class CoverCryptDecryption {
@@ -7,7 +8,13 @@ class CoverCryptDecryption {
 
   CoverCryptDecryption(this.asymmetricDecryptionKey);
 
-  Uint8List decrypt(Uint8List ciphertextBytes) {
+  CleartextHeader decryptWithAuthenticationData(
+      Uint8List ciphertextBytes, Uint8List authenticationData) {
+    return Ffi.decryptWithAuthenticationData(
+        asymmetricDecryptionKey, ciphertextBytes, authenticationData);
+  }
+
+  CleartextHeader decrypt(Uint8List ciphertextBytes) {
     return Ffi.decrypt(asymmetricDecryptionKey, ciphertextBytes);
   }
 }